refactor: 중복, 추상화 수정
- Page들로 컴포넌트 분리
- 노선 모달 2개로 (추가, 수정)분리
- statemanager대신 state로 렌더링 관리하기
