+++ index.html
@@ -13,7 +13,6 @@
           <main class="js-main mt-10 d-flex justify-center"></main>
         </div>
       </div>
-      <div class="js-modal"></div>
     </div>
   </body>
 </html>
