+++ src/css/ui/modules/modal.css
@@ -1,7 +1,5 @@
 .modal {
-  opacity: 0;
-  visibility: hidden;
-  display: flex;
+  display: none;
   position: fixed;
   top: 0;
   right: 0;
@@ -13,8 +11,7 @@
 }
 
 .modal.open {
-  opacity: 1;
-  visibility: visible;
+  display: flex;
 }
 
 .modal-inner {
