+++ src/js/core/PageComponent.js
@@ -1,0 +1,7 @@
+import Component from './Component';
+
+class PageComponent extends Component {
+  constructor() {
+    super(arguments);
+  }
+}
