--- src/js/core/StateManager.js
@@ -1,15 +1,0 @@
-class StateManager {
-  constructor() {
-    this.listeners = [];
-  }
-
-  subscribe(listener) {
-    this.listeners.push(listener);
-  }
-
-  notify() {
-    this.listeners.forEach((listener) => listener());
-  }
-}
-
-export default StateManager;
