--- src/js/stateManagers/AccessTokenManager.js
@@ -1,25 +1,0 @@
-import StateManager from '../core/StateManager.js';
-
-class AccessTokenManager extends StateManager {
-  constructor() {
-    super();
-    this.accessToken = localStorage.getItem('accessToken') || '';
-  }
-
-  setToken(token) {
-    this.accessToken = token;
-    localStorage.setItem('accessToken', this.accessToken);
-    this.notify();
-  }
-
-  getToken() {
-    return this.accessToken;
-  }
-
-  clearToken() {
-    this.setToken('');
-    localStorage.removeItem('accessToken');
-  }
-}
-
-export default AccessTokenManager;
