--- src/js/stateManagers/RouteManager.js
@@ -1,14 +1,0 @@
-import StateManager from '../core/StateManager.js';
-
-class RouteManager extends StateManager {
-  constructor() {
-    super();
-  }
-
-  goPage(route) {
-    history.pushState({ route }, null, route);
-    this.notify();
-  }
-}
-
-export default RouteManager;
